float16 -> float32

+++ detectron2/engine/defaults.py
@@ -313,7 +313,7 @@ class DefaultBatchPredictor:
                     original_image = original_image[:, :, ::-1]
                 height, width = original_image.shape[:2]
                 image = self.aug.get_transform(original_image).apply_image(original_image)
-                image = torch.as_tensor(image.astype("float16").transpose(2, 0, 1))
+                image = torch.as_tensor(image.astype("float32").transpose(2, 0, 1))
                 image.to(self.cfg.MODEL.DEVICE)
 
                 inputs.append({"image": image, "height": height, "width": width})
